Disable install button during installation process
The installToolStripMenuItem button is now disabled when clicked, during the installation process. Two lines of code were added to accomplish this: one to disable the button at the start of the installation process, and another to re-enable it after installation is complete. This is to prevent simultaneous installations which could potentially cause software failure or conflict.