Add drag-and-drop functionality for Modpack.json files
The ability to drag-and-drop Modpack.json files into the application has been introduced. Appropriate event handlers for DragEnter and DragDrop are now added to the DataGridView control. Upon a successful drop of a json file, a new configuration instance is created and existing mod details are replaced with those from the dropped file.